fixed #1 - added support for allow lists
