+++ lib/controllers/actions/oauth2.js
@@ -21,9 +21,14 @@ module.exports = function (req, res){
    * @param  {[type]} accessToken       [description]
    */
   function accessTokenResponse(error, accessToken){
-    if (error && typeof accessToken === 'undefined') {
-      return res.serverError(error);
-    } 
+    if (error || typeof accessToken === 'undefined') {
+      if(error.status) {
+        return res.status(error.status).json(error);
+      }
+      else {
+        return res.serverError(error);
+      }
+    }
     
     var attr = {
         googleEmail: accessToken.id_token.jwt.email
