+++ lib/google-oauth2.js
@@ -3,6 +3,8 @@
 var request = require('request');
 var querystring= require('querystring');
 var crypto = require('crypto');
+var _ = require('lodash');
+var GoogleOAuth2Crypto = require('./google-oauth2-crypto');
 
 exports = module.exports = GoogleOAuth2;
 
@@ -12,12 +14,15 @@ exports = module.exports = GoogleOAuth2;
  * @param {string} clientId     the google app id
  * @param {string} clientSecret the google app secret
  * @param {string} redirectURL  the url google should use as a callback
+ * @param {string} config       the google-specific configuration
+ * @param {string} allow        the allow list
  */
-function GoogleOAuth2(clientId, clientSecret, redirectURL, config){
+function GoogleOAuth2(clientId, clientSecret, redirectURL, config, allow){
   this._clientId = clientId;
   this._clientSecret = clientSecret;
   this._redirectURL = redirectURL;
   this._config = config;
+  this._allow = allow;
 }
 
 /**
@@ -93,11 +98,37 @@ GoogleOAuth2.prototype._tokenExchangeCallback = function(err, response, body){
   if(body.id_token) {
     body.id_token = this.decryptRawToken(body.id_token);
     this.accessToken = body;
-    return this._tokenCallback(null, this.accessToken);
+    var email = body.id_token.jwt.email;
+
+    if(this.authorize(email)) {
+      return this._tokenCallback(null, this.accessToken);
+    }
+    else {
+      return this._tokenCallback({status: 403, message: "Insufficient access privileges."});
+    }
   }
   else {
     return this._tokenCallback(body);
   }
+};
+
+GoogleOAuth2.prototype.authorize = function(email) {
+  if(this._allow.length === 0) {
+    return true;      // allow all
+  }
+
+  var foundMatch = false;
+  _.forEach(this._allow, function(allow) {
+    var index = allow.indexOf('@');
+    if(index <= 0 && _.endsWith(email, allow)) {
+      foundMatch = true;    // match on domain or subdomain
+    }
+    else if(email.toLowerCase() === allow.toLowerCase()) {
+      foundMatch = true;    // match on email
+    }
+  });
+
+  return foundMatch;
 };
 
 /**
@@ -162,77 +193,13 @@ GoogleOAuth2.prototype.getKey = function(kid){
       return keys[i];
     }
   }
-  return this._config.certs[kid];
+  return {};
 };
 
 GoogleOAuth2.prototype.buildPublicKey = function(key) {
-  if(key.use && key.use === 'sig') {
-    return this.rsaPublicKeyPem(key.n, key.e);
+  if(key && key.use && key.use === 'sig') {
+    var googleCrypto = new GoogleOAuth2Crypto(key.n, key.e);
+    return googleCrypto.toPemEncodedPublicKey();
   }
 };
 
-//http://stackoverflow.com/questions/18835132/xml-to-pem-in-node-js
-GoogleOAuth2.prototype.rsaPublicKeyPem = function(modulus_b64, exponent_b64) {
-  var modulus = new Buffer(modulus_b64, 'base64');
-  var exponent = new Buffer(exponent_b64, 'base64');
-
-  var modulus_hex = modulus.toString('hex');
-  var exponent_hex = exponent.toString('hex');
-
-  modulus_hex = this.prepadSigned(modulus_hex);
-  exponent_hex = this.prepadSigned(exponent_hex);
-
-  var modlen = modulus_hex.length/2;
-  var explen = exponent_hex.length/2;
-
-  var encoded_modlen = this.encodeLengthHex(modlen);
-  var encoded_explen = this.encodeLengthHex(explen);
-  var encoded_pubkey = '30' +
-      this.encodeLengthHex(
-          modlen +
-          explen +
-          encoded_modlen.length/2 +
-          encoded_explen.length/2 + 2
-      ) +
-      '02' + encoded_modlen + modulus_hex +
-      '02' + encoded_explen + exponent_hex;
-
-  var der_b64 = new Buffer(encoded_pubkey, 'hex').toString('base64');
-
-  var pem = '-----BEGIN RSA PUBLIC KEY-----\n'
-      + der_b64.match(/.{1,64}/g).join('\n')
-      + '\n-----END RSA PUBLIC KEY-----\n';
-
-  return pem;
-};
-
-GoogleOAuth2.prototype.prepadSigned = function(hexStr) {
-  var msb = hexStr[0];
-  if (msb < '0' || msb > '7') {
-      return '00'+hexStr;
-  } else {
-      return hexStr;
-  }
-};
-
-GoogleOAuth2.prototype.toHex = function(number) {
-  var nstr = number.toString(16);
-  if (nstr.length%2) {
-    return '0'+nstr;
-  }
-  return nstr;
-};
-
-// encode ASN.1 DER length field
-// if <=127, short form
-// if >=128, long form
-GoogleOAuth2.prototype.encodeLengthHex = function(n) {
-  if (n<=127) {
-    return this.toHex(n);
-  }
-  else {
-    var n_hex = this.toHex(n);
-    var length_of_length_byte = 128 + n_hex.length/2; // 0x80+numbytes
-    return this.toHex(length_of_length_byte)+n_hex;
-  }
-};
